% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{ANE}
\alias{ANE}
\title{Average distance between each accession and the nearest entry (A.NE)}
\usage{
ANE(geno.dist.all, core.names)
}
\arguments{
\item{geno.dist.all}{matrix}

\item{core.names}{vector}
}
\value{

}
\description{
Calculate distance between each accession and the nearest entry in the core and average over all accessions
}
