% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustdist.R
\name{cor.cophenetic}
\alias{cor.cophenetic}
\title{Correlate Cophenetic Distances}
\usage{
cor.cophenetic(dist.all, hc.method = "ward.D2")
}
\arguments{
\item{dist.all}{distance matrix}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}
}
\value{
correlation between cophenetic distances and original distance matrix
}
\description{
Calculates correlation between cophenetic distances and an original distance matrix.
High correlation (>0.8) can provide evidence for population structure.
}
\examples{
data(fake_geno); cor.cophenetic(dist(fake_geno))
}
