% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{cor.cophentic}
\alias{cor.cophentic}
\title{Correlate Cophenetic Distances}
\usage{
cor.cophentic(geno.dist.all, hc.method = "ward.D2")
}
\arguments{
\item{geno.dist.all}{genotypic distance matrix}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}
}
\value{

}
\description{
calculate correlation between cophenetic distances and original distance matrix
}
