% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustdist.R
\name{dist.core}
\alias{dist.core}
\title{Select a Core Collection based on a distance matrix}
\usage{
dist.core(dist.all, n.clust, size.core = 0.1, hc.method = "ward.D2")
}
\arguments{
\item{dist.all}{a distance matrix}

\item{n.clust}{number of clusters into which to group accessions}

\item{size.core}{percentage of total number of accessions to select for core, default is 10\%}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}
}
\value{
dataframe of core entries and the cluster group, pass entry column as a vector
}
\description{
Select a Core Collection based on a distance matrix
}
