% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{EE}
\alias{EE}
\title{Average distances between entries}
\usage{
EE(geno.dist.all, core.names)
}
\arguments{
\item{geno.dist.all}{matrix}

\item{core.names}{vector}
}
\value{

}
\description{
Sum distances between all pairs of entries in the core and divide by number of comparisons
}
