% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{ENE}
\alias{ENE}
\title{Average distance between each entry and nearest neighbor entry  (E.NE)}
\usage{
ENE(geno.dist.all, core.names)
}
\arguments{
\item{geno.dist.all}{matrix}

\item{core.names}{vector}
}
\value{

}
\description{
Calculate distance between nearest entries in the core and average over all entries
}
