% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_geno}
\alias{fake_geno}
\title{Genotypic data}
\format{
A data frame with 100 rows and 50 variables
}
\usage{
fake_geno
}
\description{
Dataset containing fake genotypic data at 50 markers for 100 accessions
}
\keyword{datasets}
