% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_pheno}
\alias{fake_pheno}
\title{Phenotypic and geographic data}
\format{
A data frame with 100 rows and 10 variables
}
\usage{
fake_pheno
}
\description{
Dataset containing fake phenotypic and geographic for 100 accessions
}
\keyword{datasets}
