% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\name{geno.core}
\alias{geno.core}
\title{Select a Core Collection based on genotypic distances}
\usage{
geno.core(geno.dist.all, n.clust, hc.method = "ward.D2", size.core = 0.1)
}
\arguments{
\item{geno.dist.all}{genetic distance matrix}

\item{n.clust}{number of clusters into which to group accessions}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}

\item{size.core}{percentage of total number of accessions to select for core, default is 10%}
}
\value{

}
\description{
Select a Core Collection based on genotypic distances
}
