% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{lsmeans}
\alias{lsmeans}
\title{Least Squares Means}
\usage{
lsmeans(x, id.vars, measure.vars, form, uniquecol, like, alpha = 0.05)
}
\arguments{
\item{x}{data table}

\item{id.vars}{numeric vector corresponding to columns in data}

\item{measure.vars}{numeric vector corresponding to columns in data}

\item{form}{model formula, for example "value~-1+entryID+(1|rep)"}

\item{uniquecol}{integer corresponding to entry names column- will be included in data frame output}

\item{like}{character string that indicates that something is an entry}

\item{alpha}{confidence level for confidence intervals}
}
\value{
Rdata files of lsmeans and confidence intervals
}
\description{
Calculate least squares means for experimental phenotype data
}
