% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA2dist.R
\name{MFA2dist.df}
\alias{MFA2dist.df}
\title{MFA to distance matrix}
\usage{
MFA2dist.df(mfa, n.dim, method = "euclidean")
}
\arguments{
\item{mfa}{a "MFA" object}

\item{n.dim}{numeric value specifying the number of principal components to use to construct a distance matrix}

\item{method}{the distance measure to be used. Default is "euclidean". See "stats::dist" for other options.}
}
\value{
MFA2dist.df returns a data frame of a distance object
}
\description{
This function calculates a distance matrix from a multiple factor analysis output.
}
