% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{noirot.contribution}
\alias{noirot.contribution}
\title{Noirot Principal Component scoring}
\usage{
noirot.contribution(geno.dist.all, core.names, k = 2)
}
\arguments{
\item{geno.dist.all}{genetic distance matrix}

\item{core.names}{character vector of entry names to be included in core}

\item{k}{the maximum dimensions in which to represent data; must be less than n}
}
\value{
PC score
}
\description{
Perform multidimensional scaling of genotypic distance matrix and calculate summed relative contribution of entries in core set following Noirot et al. (1996).
}
\examples{
core<-dist.core(as.matrix(dist(fake_geno)), n.clust=9)
noirot.contribution(as.matrix(dist(fake_geno)), core$name)
}
