% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{plot.geno.entries}
\alias{plot.geno.entries}
\title{MDS bi-plot of accessions}
\usage{
\method{plot}{geno.entries}(x, core.names, k = 2, axes = c(1, 2), ...)
}
\arguments{
\item{x}{genetic distance matrix}

\item{core.names}{character vector of entry names to be included in core}

\item{k}{the maximum dimensions in which to represent data; must be less than n}

\item{axes}{vector of length 2 specifying dimensions to represent with bi-plot}

\item{...}{other parameters}
}
\value{
biplot of results from multidimensional scaling of genetic distance matrix with entries selected for the core highlighted
}
\description{
MDS bi-plot of accessions
}
