% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.accuracy}
\alias{pred.accuracy}
\title{Predictive value of core entries}
\usage{
pred.accuracy(pheno, trait.col, geno.rel.all, core.names)
}
\arguments{
\item{pheno}{dataframe of phenotypic information file}

\item{trait.col}{trait column index}

\item{geno.rel.all}{genotypic  relationship matrix matrix}

\item{core.names}{character vector of core entries}
}
\value{
correlation between predicted and actual value for accessions outside of the core
}
\description{
Predictive value of core entries
}
