% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.value}
\alias{pred.value}
\title{Predictive value of core entries}
\usage{
pred.value(pheno, trait.col, geno.rel.all, core.names)
}
\arguments{
\item{pheno}{pheno file}

\item{trait.col}{trait col}

\item{geno.rel.all}{geno matrix}

\item{core.names}{core names}
}
\value{

}
\description{
Predictive value of core entries
}
