% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustdist.R
\name{cor.cophentic}
\alias{cor.cophentic}
\title{Correlate Cophenetic Distances}
\usage{
cor.cophentic(dist.all, hc.method = "ward.D2")
}
\arguments{
\item{dist.all}{distance matrix}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}
}
\value{
correlation between cophenetic distances and original distance matrix
}
\description{
calculate correlation between cophenetic distances and original distance matrix
}
