% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{EE}
\alias{EE}
\title{Average distances between entries}
\usage{
EE(geno.dist.all, core.names)
}
\arguments{
\item{geno.dist.all}{distance matrix}

\item{core.names}{character vector of entry names to be included in core}
}
\value{
average distance between entries
}
\description{
Sum distances between all pairs of entries in the core and divide by number of comparisons
}
