% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorep.R
\name{ENE}
\alias{ENE}
\title{Average distance between each entry and nearest neighbor entry  (E.NE)}
\usage{
ENE(geno.dist.all, core.names)
}
\arguments{
\item{geno.dist.all}{genetic distance matrix}

\item{core.names}{character vector of entry names to be included in core}
}
\value{
average distance between each entry and nearest neighbor entry
}
\description{
Calculate distance between nearest entries in the core and average over all entries. Value should be as large as possible (see Odong et al. 2013).
}
\examples{
core<-dist.core(as.matrix(dist(fake_geno)), n.clust=9); ENE(dist(fake_geno), core$name)
}
