% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{lsmeans}
\alias{lsmeans}
\title{Least Squares Means}
\usage{
lsmeans(x, id.vars, measure.vars, form, uniquecol, out = "means", alpha = NULL)
}
\arguments{
\item{x}{a data frame containing entry names,}

\item{id.vars}{numeric vector corresponding to ID variables columns in data}

\item{measure.vars}{numeric vector corresponding to measured variables columns in data}

\item{form}{model formula, for example "value~-1+entryID+(1|rep)"}

\item{uniquecol}{integer corresponding to entry names column- will be included in data frame output}

\item{out}{data to return. Should be either "means" for least-squares means estimates or "ci" for confidence intervals.}

\item{alpha}{significance level used to construct confidence intervals. Only used when out="ci"}
}
\value{
dataframe of means or confidence intervals
}
\description{
Calculate least squares means for experimental phenotype data
}
