% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustdist.R
\name{sil.width}
\alias{sil.width}
\title{Calculate Average Silhouette Widths}
\usage{
sil.width(
  dist.all,
  n.clust.to.check = 9,
  hc.func = "hclust",
  hc.method = "ward.D2",
  plot.sil = TRUE
)
}
\arguments{
\item{dist.all}{distance matrix}

\item{n.clust.to.check}{largest number of clusters to test, must be less than number of accessions}

\item{hc.func}{hierarchical clustering function to be used. Default is "hclust". See factoextra::hcut for other options.}

\item{hc.method}{agglomeration method to be used. Default method is "ward.D2" See stats::hclust for other options.}

\item{plot.sil}{TRUE/FALSE to plot silhouette widths}
}
\value{
data frame of average silhouette width for different numbers of clusters

scree plot of average silhouette widths
}
\description{
Calculate Average Silhouette Widths
}
